requirements.txt
keyring
pyyaml
requests >= 2.4.2
